Remove locales from main repository
Extract another repository only concerned with locales:
https://github.com/thoughtbot/paperclip-i18n

A user of paperclip might choose to add `paperclip-i18n` as a dependency
for their application, or copy over the related locale file into their
project.

This is a first step in splitting some parts of the codebase into other
projects. Simple and relevant contributions like new translations might
get lost in the shuffle of dozens of PRs awaiting for approval, that
take time to review and merge.

Locales are a simple enough problem to be handled in an external
repository, where tests for consistency across translations run in less
than a second, and PRs can almost trivially be merged in.

Other README adjustments.